Cleanup of Folder related Rules
Folder Pools are now Acting as normal Folders. Only open is how to reset them. But this is WIP.

Basefolder Rules are now obsolte, the normal move_folder leads to the same result cause of the last change
of folder Handling. This is marked now in the GUI and using the old ones will raise an exception