New Feature: Objects
Objects are a new type for the CMDB Syncer. Set "IS Object" in an
account, and the hosts is importet as an Object. Objects can have
labels and attributes like hosts, which can be used everywhere, but
they are not exportetd like Hosts.

Also when creating Checkmk Groups, the Object Name can be used as Group
name now. Use the foreach type "Object" for that

Developers:
If you have a own export plugin,
Please migrate to Host.get_export_host() from using Hosts.ojbects().
Otherwise you will export objects in the future (of course only if you
import some)